P2P messaging refactor and history debug

+++ app/P2P/messages/messages.go
@@ -5,12 +5,12 @@ type P2PClientMessage struct {
 	PeerID    string `json:"peer_id"`
 	MessageID uint64 `json:"message_id"`
 	Method    string `json:"method"`
-	Payload   []byte `json:"payload"`
+	Payload   []int  `json:"payload"`
 }
 
 // server message
 type P2PServerMessasge struct {
 	MessageID int    `json:"message_id"`
 	Method    string `json:"method"`
-	Payload   []byte `json:"payload"`
+	Payload   []int  `json:"payload"`
 }
